[REFACTOR] : 폴더 구조 변경 - shared 폴더를 components 폴더 안으로 이동
